+++ src/test/java/nextstep/subway/unit/LineServiceMockTest.java
@@ -6,10 +6,10 @@ import nextstep.subway.applicaion.dto.LineRequest;
 import nextstep.subway.applicaion.dto.LineResponse;
 import nextstep.subway.applicaion.dto.LineUpdateRequest;
 import nextstep.subway.applicaion.dto.SectionRequest;
-import nextstep.subway.domain.Line;
-import nextstep.subway.domain.LineRepository;
-import nextstep.subway.domain.Section;
-import nextstep.subway.domain.Station;
+import nextstep.subway.domain.line.Line;
+import nextstep.subway.domain.line.LineRepository;
+import nextstep.subway.domain.section.Section;
+import nextstep.subway.domain.station.Station;
 import nextstep.subway.exception.NotFoundException;
 import nextstep.subway.exception.advice.ValidationException;
 import org.junit.jupiter.api.BeforeEach;
